+++ src/hooks/useCartSelector.ts
@@ -1,6 +1,7 @@
 import { useSelector } from 'react-redux';
 import { RootState } from '@/redux/store';
-import { CartItem } from '@/redux/modules/cart';
+
+import { CartItem } from '@/types';
 
 const selectCartItemById = (state: RootState, targetId: number) => {
   return state.cart.items.find((item) => item.id === targetId);
